requirements.txt
redis==2.10.6
tornado==4.5.1
pycrypto==2.6.1
